Test environment consistency fixes
git-clean in E107Base now runs first in _beforeSuite() so that tainted
files are not uploaded in Acceptance tests

E107Base::revokeLocalE107Config() now checks for file existence before
unlinking. XXX: Could remove this method if cleanVCS() proves reliable

Acceptance won't try to write an e107_config.php because the tests
themselves populate the configuration file

+++ tests/_support/Helper/E107Base.php
@@ -11,8 +11,8 @@ abstract class E107Base extends Base
 
 	public function _beforeSuite($settings = array())
 	{
-		parent::_beforeSuite($settings);
 		$this->cleanVCS();
+		parent::_beforeSuite($settings);
 		$this->writeLocalE107Config();
 	}
 
@@ -45,7 +45,8 @@ abstract class E107Base extends Base
 
 	protected function revokeLocalE107Config()
 	{
-		unlink(self::APP_PATH_E107_CONFIG);
+		if (file_exists(self::APP_PATH_E107_CONFIG))
+			unlink(self::APP_PATH_E107_CONFIG);
 	}
 
 	protected function cleanVCS()
